// Boardlight/src/api/generated/models/shopIndexQRCodesResponse.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Blueboard
 * OpenAPI spec version: 4.1.0
 */

export interface ShopIndexQRCodesResponse {
  createdAt?: string;
  /** @nullable */
  email?: string | null;
  id?: number;
  /** @nullable */
  name?: string | null;
  updatedAt?: string;
}